+++ packages/twoslash/CHANGELOG.md
@@ -1,5 +1,11 @@
 # fumadocs-twoslash
 
+## 1.0.2
+
+### Patch Changes
+
+- 08e4904: Update types
+
 ## 1.0.1
 
 ### Patch Changes